[Shows] Rename ShowsMetadata because its in the shows package anyways.
