+++ lib/components/artist/shows/large_list.js
@@ -8,7 +8,7 @@ const { View, Text, StyleSheet, ListView } = React;
 import Headline from '../../text/headline'
 import SerifText from '../../text/serif'
 import OpaqueImageView from '../../opaque_image_view'
-import ShowMetadata from './show_metadata'
+import ShowMetadata from './metadata'
 
 
 class LargeList extends React.Component {
