+++ lib/components/artist/shows/metadata.js
@@ -7,7 +7,7 @@ const { View, Text, StyleSheet } = React;
 
 import SerifText from '../../text/serif'
 
-class ShowMetadata extends React.Component {
+class Metadata extends React.Component {
 
   render() {
     return (
@@ -49,7 +49,7 @@ const styles = StyleSheet.create({
   },
 });
 
-export default Relay.createContainer(ShowMetadata, {
+export default Relay.createContainer(Metadata, {
   fragments: {
     show: () => Relay.QL`
       fragment on PartnerShow {
