+++ lib/components/artist/shows/small_list.js
@@ -8,7 +8,7 @@ const { View, Text, StyleSheet, ListView } = React;
 import Headline from '../../text/headline'
 import SerifText from '../../text/serif'
 import OpaqueImageView from '../../opaque_image_view'
-import ShowMetadata from './show_metadata'
+import ShowMetadata from './metadata'
 
 
 class SmallList extends React.Component {
